cambios soto

+++ tateiv2.sql
@@ -320,7 +320,7 @@ DROP TABLE IF EXISTS perfil;
 		id_agenda_detalle INTEGER NOT NULL AUTO_INCREMENT,
 		id_agenda INTEGER NOT NULL,
 		fecha DATE NOT NULL,
-		avance_esperado INTEGER NOT NULL,
+		avance_esperado FLOAT NOT NULL,
 		observaciones VARCHAR(150) NULL,		
 		fecha_alta DATETIME NOT NULL,
 		usuario_alta VARCHAR(20) NOT NULL,
